refactor: clean up provider typing
